@@ -9,11 +9,12 @@
         }
     ],
     "require": {
-        "symfony/framework-bundle": "^5.2",
-        "doctrine/doctrine-bundle": "^2.3",
-        "doctrine/orm": "^2.8",
-        "symfony/uid": "^5.2",
-        "symfony/doctrine-bridge": "^5.2"
+        "symfony/framework-bundle": "^8.0",
+        "doctrine/doctrine-bundle": "^3.1",
+        "doctrine/orm": "^3.5",
+        "symfony/uid": "^8.0",
+        "symfony/doctrine-bridge": "^8.0",
+        "symfony/yaml": "^8.0"
     },
     "autoload": {
         "psr-4": {
@@ -26,6 +27,7 @@
         }
     },
     "require-dev": {
-        "doctrine/doctrine-fixtures-bundle": "^3.4"
+        "doctrine/doctrine-fixtures-bundle": "^4.3",
+        "phpunit/phpunit": "^12.5"
     }
 }
