@@ -18,7 +18,7 @@ use Symfony\Component\Config\FileLocator;
  */
 class WilayahExtension extends Extension
 {
-    public function load(array $configs, ContainerBuilder $container) 
+    public function load(array $configs, ContainerBuilder $container) :void
     {
         $configuration = new Configuration();
         $config = $this->processConfiguration($configuration, $configs);
